refactor: extract validate_cluster into controller::validate module
